requirements.txt
Django==5.0
djangorestframework==3.14.0
pandas==2.1.4
celery==5.3.6